Create fragmentsGeneratory.py
I wrote this code to generate packets destined to the vRouter which gets transmitted out-of-order. 10.204.75.3 is the floating IP of a virtual-machine running on compute. This is what happens when vRouter receives fragments that are out of order. 

Suppose, vRouter receives in this sequence F3, F4, F5, F0,  F1 and F2. 

In the above example till F0 is received we can’t create the Flow as we don’t have complete 5 tuple . So in this case F3, F4, F5 is held in Assembler till F0 is received. This is nothing to do with Flow hold state. Once F0 is received we create the Flow entry and hold the fragments F0, F3 and F4 by marking flow entry as hold till agent approves the flow.